clo/requirements.txt
django==4.2.3
djangorestframework==3.14.0
django_filter==23.3
django-cors-headers==4.3.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
psycopg2_binary==2.9.9
gunicorn==21.2.0
django-scheduler==0.10.1
django-bower==5.2.0
